requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
requests>=2.28.0
yfinance>=0.2.18
python-dotenv>=1.0.0
jupyter>=1.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0